requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-ollama
langgraph
pydantic
groq
pinecone-client
ollama
langchain-text-splitters 